Refactor ファイルを役割ごとに分割
