+++ 05.ls/ls_modules/long_format.rb
@@ -1,105 +1,14 @@
 # frozen_string_literal: true
 
-require 'byebug'
-require 'optparse'
+require_relative 'constant'
+require_relative 'default_format'
 require 'etc'
 require 'date'
 
-class Option
-  attr_reader :options
-
-  def initialize
-    @options = {}
-    OptionParser.new do |option|
-      option.on('-a') { |v| @options[:select_all_files] = v }
-      option.on('-l') { |v| @options[:long_format] = v }
-      option.on('-r') { |v| @options[:reverse_sort] = v }
-      option.parse!(ARGV)
-    end
-  end
-end
-
 module ListSegment
-  NO_FILE_OPTION = 0
-
-  PERMISSION_PATTERNS = {
-    '0' => '---',
-    '1' => '--x',
-    '2' => '-w-',
-    '3' => '-wx',
-    '4' => 'r--',
-    '5' => 'r-x',
-    '6' => 'rw-',
-    '7' => 'rwx'
-  }.freeze
-
-  UID_PERMISSION_PATTERNS = {
-    '0' => '--S',
-    '1' => '--s',
-    '2' => '-wS',
-    '3' => '-ws',
-    '4' => 'r-S',
-    '5' => 'r-s',
-    '6' => 'rwS',
-    '7' => 'rws'
-  }.freeze
-
-  STICKY_PERMISSION_PATTERNS = {
-    '0' => '--T',
-    '1' => '--t',
-    '2' => '-wT',
-    '3' => '-wt',
-    '4' => 'r-T',
-    '5' => 'r-t',
-    '6' => 'rwT',
-    '7' => 'rwt'
-  }.freeze
-
-  class DefaultFormat
-    def initialize(options = {}, column_num = 3)
-      @options = options
-      @column_num = column_num
-      @files = sort_files(Dir.glob('*', to_fnm))
-      @stats = to_stats(@files) if options[:long_format]
-    end
-
-    def output
-      row_num = calc_row_num
-      row_num.times do |row|
-        @column_num.times do |column|
-          file = @files[column * row_num + row]
-          break if file.nil?
-
-          print file.to_s.ljust(count_max_file_name_str).to_s
-        end
-        print "\n"
-      end
-    end
-
-    private
-
-    def to_fnm
-      @options[:select_all_files] ? File::FNM_DOTMATCH : NO_FILE_OPTION
-    end
-
-    def sort_files(files)
-      @options[:reverse_sort] ? files.sort.reverse : files.sort
-    end
-
-    def mod
-      @files.size % @column_num
-    end
-
-    def calc_row_num
-      (@files.size / @column_num) + mod
-    end
-
-    def count_max_file_name_str(add_space = 2)
-      @files.map(&:length).max + add_space
-    end
-  end
-
   class LongFormat < DefaultFormat
+    include LongFormatConstant
+
     def initialize(options = {})
       super
       @stats = to_stats(@files)
@@ -194,7 +103,3 @@ module ListSegment
     end
   end
 end
-
-opt = Option.new
-ls = opt.options[:long_format] ? ListSegment::LongFormat.new(opt.options) : ListSegment::DefaultFormat.new(opt.options)
-ls.output
